Movement, type updates

+++ engine/game/damage.cpp
@@ -29,25 +29,25 @@ void DamageReceiver::ApplyDamage(const DamageEvent& evt)
 	}
 }
 
-float DamageCalculator::CalculateArmorResistPercent(float enemy_level, float armor)
+double DamageCalculator::CalculateArmorResistPercent(double enemy_level, double armor)
 {
-	const float BaseArmorResist = 5;
+	const double BaseArmorResist = 5;
 
-	const float enemy_level_ratio = 5 * enemy_level;
-	const float armor_ratio = 100;
+	const double enemy_level_ratio = 5 * enemy_level;
+	const double armor_ratio = 100;
 
-	float DR = armor / (enemy_level + armor_ratio + armor) * 100;
+	double DR = armor / (enemy_level + armor_ratio + armor) * 100;
 	return DR;
 }
 
-float DamageCalculator::CalculateResistPercent(float enemy_level, float resist)
+double DamageCalculator::CalculateResistPercent(double enemy_level, double resist)
 {
-	const float BaseArmorResist = 0;
+	const double BaseArmorResist = 0;
 
-	const float enemy_level_ratio = 5 * enemy_level;
-	const float mr_ratio = 50;
+	const double enemy_level_ratio = 5 * enemy_level;
+	const double mr_ratio = 50;
 
-	float DR = resist / (enemy_level + mr_ratio + resist) * 100;
+	double DR = resist / (enemy_level + mr_ratio + resist) * 100;
 	return DR;
 }
